[#166] refactor: 서비스 페이지 폴더구조 변경
