+++ src/app/(route)/(main)/service/privacy/page.tsx
@@ -1,6 +1,5 @@
-import TermsGuide from '@/app/_components/service/TermsGuide';
 import { PrivacyData } from '@/app/_constant/service/terms';
-import React from 'react';
+import { TermsGuide } from '@/app/_features/service';
 
 const Page = () => {
   return <TermsGuide title="개인정보처리방침" data={PrivacyData} />;
