+++ src/app/(route)/(main)/service/tos/page.tsx
@@ -1,6 +1,5 @@
-import TermsGuide from '@/app/_components/service/TermsGuide';
 import { TermsData } from '@/app/_constant/service/terms';
-import React from 'react';
+import { TermsGuide } from '@/app/_features/service';
 
 const Page = () => {
   return <TermsGuide title="이용약관" data={TermsData} />;
