+++ src/app/_features/service/index.ts
@@ -1,0 +1,1 @@
+export { default as TermsGuide } from './components/TermsGuide';
