+++ src/app/_features/service/types.ts
@@ -1,0 +1,10 @@
+export interface TermsGuideProsp {
+  title: string;
+  data: {
+    section: number;
+    title: string;
+    list: {
+      content: string;
+    }[];
+  }[];
+}
